refactor: reduce size of files in core by splitting into relevant sub-files
