+++ Sources/WebUI/Core/Website.swift
@@ -1,104 +1,6 @@
 import Foundation
 import Logging
 
-/// Represents a rule in a robots.txt file.
-///
-/// Used to define instructions for web crawlers about which parts of the website should be crawled.
-public struct RobotsRule {
-  /// The user agent the rule applies to (e.g., "Googlebot" or "*" for all crawlers).
-  public let userAgent: String
-
-  /// Paths that should not be crawled.
-  public let disallow: [String]?
-
-  /// Paths that are allowed to be crawled (overrides disallow rules).
-  public let allow: [String]?
-
-  /// The delay between successive crawls in seconds.
-  public let crawlDelay: Int?
-
-  /// Creates a new robots.txt rule.
-  ///
-  /// - Parameters:
-  ///   - userAgent: The user agent the rule applies to (e.g., "Googlebot" or "*" for all crawlers).
-  ///   - disallow: Paths that should not be crawled.
-  ///   - allow: Paths that are allowed to be crawled (overrides disallow rules).
-  ///   - crawlDelay: The delay between successive crawls in seconds.
-  ///
-  /// - Example:
-  ///   ```swift
-  ///   let rule = RobotsRule(
-  ///     userAgent: "*",
-  ///     disallow: ["/admin/", "/private/"],
-  ///     allow: ["/public/"],
-  ///     crawlDelay: 10
-  ///   )
-  ///   ```
-  public init(
-    userAgent: String,
-    disallow: [String]? = nil,
-    allow: [String]? = nil,
-    crawlDelay: Int? = nil
-  ) {
-    self.userAgent = userAgent
-    self.disallow = disallow
-    self.allow = allow
-    self.crawlDelay = crawlDelay
-  }
-}
-
-/// Represents a sitemap entry for a URL with metadata.
-///
-/// Used to define information for a URL to be included in a sitemap.xml file.
-/// Follows the Sitemap protocol as defined at https://www.sitemaps.org/protocol.html.
-public struct SitemapEntry {
-  /// The URL of the page.
-  public let url: String
-
-  /// The date when the page was last modified.
-  public let lastModified: Date?
-
-  /// The expected frequency of changes to the page.
-  public enum ChangeFrequency: String {
-    case always, hourly, daily, weekly, monthly, yearly, never
-  }
-
-  /// How frequently the page is likely to change.
-  public let changeFrequency: ChangeFrequency?
-
-  /// The priority of this URL relative to other URLs on your site (0.0 to 1.0).
-  public let priority: Double?
-
-  /// Creates a new sitemap entry for a URL.
-  ///
-  /// - Parameters:
-  ///   - url: The URL of the page.
-  ///   - lastModified: The date when the page was last modified.
-  ///   - changeFrequency: How frequently the page is likely to change.
-  ///   - priority: The priority of this URL relative to other URLs (0.0 to 1.0).
-  ///
-  /// - Example:
-  ///   ```swift
-  ///   let homepage = SitemapEntry(
-  ///     url: "https://example.com/",
-  ///     lastModified: Date(),
-  ///     changeFrequency: .weekly,
-  ///     priority: 1.0
-  ///   )
-  ///   ```
-  public init(
-    url: String,
-    lastModified: Date? = nil,
-    changeFrequency: ChangeFrequency? = nil,
-    priority: Double? = nil
-  ) {
-    self.url = url
-    self.lastModified = lastModified
-    self.changeFrequency = changeFrequency
-    self.priority = priority
-  }
-}
-
 /// Manages and builds a collection of routes (documents) into a complete static website.
 ///
 /// The `Website` struct serves as the entry point for creating a static site, managing
